Login page
Added function to find esisting users.

When login in, the page can detect exsisting users. The login page uses localStorage to store a list of users

if it can find an exsisting user from localStorage, it will print out the users email on the page

if not, the page will tell it couldn't find the user

+++ login/js/login.js
@@ -1,5 +1,20 @@
 const emailField = document.getElementById("email-field");
 const passwordField = document.getElementById("password-field");
+const activeUserDisplay = document.getElementById("active-user-display");
+let activeUser = JSON.parse(localStorage.getItem("user"));
+
+let userList = [];
+const userData = JSON.parse(localStorage.getItem("userList"))
+
+if(userData){
+    userList = [...userData]
+}
+
+if(activeUser){
+    activeUserDisplay.innerText = `Welcome ${activeUser.email}!`;
+}
+
+console.log(userList);
 
 document.getElementById("login-button").addEventListener("click", e =>{
     e.preventDefault();
@@ -13,8 +28,29 @@ document.getElementById("login-button").addEventListener("click", e =>{
         password: passwordField.value
     }
 
-    localStorage.setItem("user", JSON.stringify(loginInfo));
-    emailField.value = "";
-    passwordField.value = "";
-    console.log(JSON.parse(localStorage.getItem("user")));
+    
+    if(isExistingUser(loginInfo)){
+        localStorage.setItem("user", JSON.stringify(loginInfo));
+        activeUser = loginInfo;
+        activeUserDisplay.innerText = `Welcome ${activeUser.email}!`;
+        emailField.value = "";
+        passwordField.value = "";
+    }
+    else{
+        activeUserDisplay.innerText = `User doesn't exist`;
+        passwordField.value = "";
+    }
 })
+
+const isExistingUser = (loginDetail) => {
+    console.log(`loginDetail: ${loginDetail.email} ${loginDetail.password}`);
+    let flag = false;
+    userList.forEach(user =>{
+        console.log(`user: ${user.email} ${user.password}`);
+        if(user.email === loginDetail.email && user.password === loginDetail.password){
+            console.log("found user!");
+            flag = true;
+        }
+    })
+    return flag;
+}
